Prepare for 64-bit positions
This does not add 64-bit position support, it just ensures samtools
will build without warnings (using gcc/clang -Wall) when built
against both 32-bit and 64-bit position branches of HTSlib.

Makes various printf()s use PRId64 for positions.  The values
are cast to (int64_t) so both 32-bit and 64-bit versions work.

Fix isize calculation in bam_mate.c so it sign-extends correctly.

+++ test/merge/test_bam_translate.c
@@ -31,10 +31,11 @@ DEALINGS IN THE SOFTWARE.  */
 #include <string.h>
 #include <errno.h>
 #include <unistd.h>
+#include <inttypes.h>
 
 void dump_read(bam1_t* b) {
     printf("->core.tid:(%d)\n", b->core.tid);
-    printf("->core.pos:(%d)\n", b->core.pos);
+    printf("->core.pos:(%"PRId64")\n", (int64_t) b->core.pos);
     printf("->core.bin:(%d)\n", b->core.bin);
     printf("->core.qual:(%d)\n", b->core.qual);
     printf("->core.l_qname:(%d)\n", b->core.l_qname);
@@ -42,8 +43,8 @@ void dump_read(bam1_t* b) {
     printf("->core.n_cigar:(%d)\n", b->core.n_cigar);
     printf("->core.l_qseq:(%d)\n", b->core.l_qseq);
     printf("->core.mtid:(%d)\n", b->core.mtid);
-    printf("->core.mpos:(%d)\n", b->core.mpos);
-    printf("->core.isize:(%d)\n", b->core.isize);
+    printf("->core.mpos:(%"PRId64")\n", (int64_t) b->core.mpos);
+    printf("->core.isize:(%"PRId64")\n", (int64_t) b->core.isize);
     if (b->data) {
         printf("->data:");
         int i;
